Consolidate DiDeMo retrieval tasks into a single file
Merge 4 separate DiDeMo task files into didemo_retrieval.py with a
shared _load_didemo helper, reducing duplication while preserving
all task names and metadata.

Made-with: Cursor
